# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/colors": {
      "post": {
        "operationId": "ColorsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateColorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Colors"
        ]
      },
      "get": {
        "operationId": "ColorsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Colors"
        ]
      }
    },
    "/colors/{id}": {
      "get": {
        "operationId": "ColorsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Colors"
        ]
      },
      "patch": {
        "operationId": "ColorsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateColorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Colors"
        ]
      },
      "delete": {
        "operationId": "ColorsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Colors"
        ]
      }
    },
...
    "/collections": {
      "post": {
        "operationId": "CollectionsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCollectionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Collection created successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "409": {
            "description": "Slug already exists"
          }
        },
        "summary": "Create a new collection",
        "tags": [
          "collections"
        ]
      },
      "get": {
        "operationId": "CollectionsController_findAll",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Cursor token for pagination",
            "schema": {
              "example": "eyJpZCI6IjEyMyIsImNyZWF0ZWRfYXQiOiIyMDI0LTAxLTAxVDAwOjAwOjAwLjAwMFoifQ==",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Collections retrieved successfully"
          }
        },
        "summary": "Get all collections with cursor pagination",
        "tags": [
          "collections"
        ]
      }
    },
    "/collections/{id}": {
      "get": {
        "operationId": "CollectionsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Collection UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Collection found"
          },
          "404": {
            "description": "Collection not found"
          }
        },
        "summary": "Get collection by ID",
        "tags": [
          "collections"
        ]
      },
      "patch": {
        "operationId": "CollectionsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Collection UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCollectionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Collection updated successfully"
          },
          "404": {
            "description": "Collection not found"
          },
          "409": {
            "description": "Slug already exists"
          }
        },
        "summary": "Update collection",
        "tags": [
          "collections"
        ]
      },
      "delete": {
        "operationId": "CollectionsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Collection UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Collection deleted successfully"
          },
          "404": {
            "description": "Collection not found"
          }
        },
        "summary": "Delete collection",
        "tags": [
          "collections"
        ]
      }
    },
    "/collections/slug/{slug}": {
      "get": {
        "operationId": "CollectionsController_findBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "description": "Collection slug",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Collection found"
          },
          "404": {
            "description": "Collection not found"
          }
        },
        "summary": "Get collection by slug",
        "tags": [
          "collections"
        ]
      }
    },
    "/collections/{id}/products": {
      "get": {
        "operationId": "CollectionsController_getProducts",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Collection UUID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Cursor token for pagination",
            "schema": {
              "example": "eyJpZCI6IjEyMyIsImNyZWF0ZWRfYXQiOiIyMDI0LTAxLTAxVDAwOjAwOjAwLjAwMFoifQ==",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Products retrieved successfully"
          },
          "404": {
            "description": "Collection not found"
          }
        },
        "summary": "Get products in a collection with cursor pagination",
        "tags": [
          "collections"
        ]
      },
      "post": {
        "operationId": "CollectionsController_assignProducts",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Collection UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignProductsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Products assigned successfully"
          },
          "400": {
            "description": "Invalid product IDs"
          },
          "404": {
            "description": "Collection or products not found"
          }
        },
        "summary": "Assign products to a collection",
        "tags": [
          "collections"
        ]
      },
      "delete": {
        "operationId": "CollectionsController_removeProducts",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Collection UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignProductsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Products removed successfully"
          },
          "404": {
            "description": "Collection not found"
          }
        },
        "summary": "Remove products from a collection",
        "tags": [
          "collections"
        ]
      }
    },
    "/collections/{id}/products/count": {
      "get": {
        "operationId": "CollectionsController_getProductCount",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Collection UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product count retrieved"
          },
          "404": {
            "description": "Collection not found"
          }
        },
        "summary": "Get count of products in a collection",
        "tags": [
          "collections"
        ]
      }
    }
  },
  "info": {
    "title": "Ecom API",
    "description": "E-commerce REST API documentation",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateColorDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateColorDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateCollectionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Summer Collection 2024",
            "description": "Collection name"
          },
          "slug": {
            "type": "string",
            "example": "summer-collection-2024",
            "description": "Collection slug (auto-generated if not provided)"
          },
          "description": {
            "type": "string",
            "example": "Discover our latest summer fashion collection",
            "description": "Collection description"
          },
          "banner_image_url": {
            "type": "string",
            "example": "https://example.com/banner.jpg",
            "description": "Banner image URL"
          },
          "seo_title": {
            "type": "string",
            "example": "Summer Collection 2024 - Fashion Store",
            "description": "SEO title"
          },
          "seo_description": {
            "type": "string",
            "example": "Browse our curated summer collection with the latest trends",
            "description": "SEO description"
          },
          "is_active": {
            "type": "boolean",
            "example": true,
            "description": "Is collection active",
            "default": true
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateCollectionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Summer Collection 2024",
            "description": "Collection name"
          },
          "slug": {
            "type": "string",
            "example": "summer-collection-2024",
            "description": "Collection slug (auto-generated if not provided)"
          },
          "description": {
            "type": "string",
            "example": "Discover our latest summer fashion collection",
            "description": "Collection description"
          },
          "banner_image_url": {
            "type": "string",
            "example": "https://example.com/banner.jpg",
            "description": "Banner image URL"
          },
          "seo_title": {
            "type": "string",
            "example": "Summer Collection 2024 - Fashion Store",
            "description": "SEO title"
          },
          "seo_description": {
            "type": "string",
            "example": "Browse our curated summer collection with the latest trends",
            "description": "SEO description"
          },
          "is_active": {
            "type": "boolean",
            "example": true,
            "description": "Is collection active",
            "default": true
          }
        }
      },
      "AssignProductsDto": {
        "type": "object",
        "properties": {
          "productIds": {
            "example": [
              "550e8400-e29b-41d4-a716-446655440000",
              "660e8400-e29b-41d4-a716-446655440001"
            ],
            "description": "Array of product IDs to assign to the collection",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "productIds"
        ]
      }
    }
  }
}